Refactor: Simplify closures and replace manual loops with idiomatic iterators
Remove unnecessary closures on `Lazy::new` and `unwrap_or_else`, replace
`while let` loops with `for`, consolidate `Default` implementations, and
apply Clippy lint suppressions where needed.

+++ src/api/graphql/router.rs
@@ -39,7 +39,7 @@ async fn execute_graphql(
         ));
     }
 
-    let compiler = ASTCompiler::new(config.graphql.max_query_depth as i32);
+    let compiler = ASTCompiler::new(config.graphql.max_query_depth);
     let operations = compiler.compile(&payload.query).map_err(|e| {
         AxiomError::new(
             "GRAPHQL_COMPILE_ERROR",
